update altair-fastify-plugin example
formerly called "fastify-gql" was renamed to "mercurius"
