+++ packages/altair-fastify-plugin/package.json
@@ -26,20 +26,20 @@
     "example"
   ],
   "scripts": {
-    "example": "ts-node example/fastifyGQL.ts",
+    "example": "ts-node example/mercurius.ts",
     "prepare": "tsc",
     "test": "echo \"Error: no test specified\" && exit 0"
   },
   "dependencies": {
     "altair-static": "^2.5.1",
-    "fastify-plugin": "^2.0.1",
-    "fastify-static": "^3.2.0"
+    "fastify-plugin": "^2.3.4",
+    "fastify-static": "^3.2.1"
   },
   "devDependencies": {
-    "fastify": "^3.1.1",
-    "fastify-gql": "^5.1.1",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.9.7"
+    "fastify": "^3.5.1",
+    "mercurius": "^6.0.0",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.3"
   },
   "peerDependencies": {
     "fastify": "^3.1.1"
